Macro aoc::localpath

macro_rules! localpath {
    ($tt:tt) => { ... };
}

Return a path to the input, starting at CARGO_MANIFEST_DIR of the caller, in a way that compiles the result into the final binary.